requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.12.0
python-dotenv==1.0.0
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
ics==0.7.2
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
langchain[all]>=0.1.0
langchain-ollama>=0.1.0  # 用于 Ollama LLM 集成
langchain-huggingface>=0.0.1  # 用于 HuggingFace 嵌入模型
faiss-cpu>=1.7.4
tiktoken>=0.5.0
sentence-transformers>=2.2.0

